Refactor map component, run geocoder on server
Change-type: minor
