+++ lib/build-runner.js
@@ -133,6 +133,7 @@ exports.run = async ({
       LANDR_OUTPUT_DIRECTORY: outputDir,
       LANDR_DEPLOY_URL: siteOptions.url,
       LANDR_GOOGLE_MAPS_KEY: process.env.LANDR_GOOGLE_MAPS_KEY,
+      LANDR_GOOGLE_GEOCODE_KEY: process.env.LANDR_GOOGLE_GEOCODE_KEY,
       LANDR_MIXPANEL_TOKEN: process.env.LANDR_MIXPANEL_TOKEN,
       LANDR_MIXPANEL_PROXY: process.env.LANDR_MIXPANEL_PROXY,
       LANDR_THEME: JSON.stringify(siteTheme),
