+++ lib/cli/utils.js
@@ -27,7 +27,7 @@ exports.getMetaData = async (repository, branch) => {
     reference: branch
   })
 
-  const contract = metaGenerator.run(scrutinizerData)
+  const contract = await metaGenerator.run(scrutinizerData)
 
   return contract
 }
